// Ejercicios Angular/tareas/src/app/tareas/search-on-type/interfaces/universidad.interface.ts
// Generated by https://quicktype.io

export interface Universidades {
    domains:          string[];
    alpha_two_code:   AlphaTwoCode;
    country:          Country;
    web_pages:        string[];
    name:             string;
    "state-province": null | string;
}

export enum AlphaTwoCode {
    Es = "ES",
}

export enum Country {
    Spain = "Spain",
}